+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/coltiebaby/g-law
+module github.com/coltiebaby/glaw
 
 go 1.12
