+++ v3/champions/api.go
@@ -1,8 +1,8 @@
 package champions
 
 import (
-	"github.com/coltiebaby/g-law/riot"
-	"github.com/coltiebaby/g-law/riot/v3"
+	"github.com/coltiebaby/glaw/"
+	"github.com/coltiebaby/glaw/v3"
 )
 
 var buildUri = v3.BuildUriFunc(`platform`)
