+++ v4/matches/api.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"net/url"
 
-	"github.com/coltiebaby/g-law/riot"
-	"github.com/coltiebaby/g-law/riot/v4"
+	"github.com/coltiebaby/glaw/riot"
+	"github.com/coltiebaby/glaw/riot/v4"
 )
 
 var buildUri = v4.BuildUriFunc(`match`)
